+++ CHANGELOG.md
@@ -20,3 +20,5 @@
 - removed `kindOf` from `ToolTip`
 - added `isLight` to `ToolTip`. The default now has dark styling, and this adds light styling.
 - added `isVariableCountWidth` to `Tabs`. Allows count badges to have fixed or variable widths.
+- added `isIncludeExclude` to `Switch`. Adds red/green styling.
+